+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.1] - 2021-01-25
 
 - Readme updated
+- Add info if update of CLI is available
 
 ## [0.1.0] - 2021-01-22