+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Simpleen is a [translation platform for i18n projects](https://simpleen.io/).
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2021-02-01
+
+- Add new languages supported by google translate
+
 ## [0.2.0] - 2021-01-26
 
 - New options for interpolations, run simpleen init to use them